+++ lib/clixon/clixon_xml.h
@@ -56,6 +56,11 @@
 
 /* Collections namespace from draft-ietf-netconf-restconf-collection-00.txt
  */
+#define NETCONF_COLLECTION_NAMESPACE "urn:ietf:params:xml:ns:yang:ietf-netconf-collection"
+
+/* See RFC 7950 Sec 5.3.1: YANG defines an XML namespace for NETCONF <edit-config> 
+ * operations, <error-info> content, and the <action> element.
+ */
 #define NETCONF_COLLECTION_NAMESPACE "urn:ietf:params:xml:ns:yang:ietf-netconf-list-pagination"
 
 /* Output symbol for netconf get/get-config