+++ lib/src/clixon_netconf_lib.c
@@ -1516,7 +1516,6 @@ netconf_module_load(clicon_handle h)
     /* Load restconf yang. Note this is also a part of clixon-config */
     if (yang_spec_parse_module(h, "clixon-restconf", NULL, yspec)< 0)
 	goto done;
-#if 1
     /* XXX: Both the following settings are because clicon-handle is not part of all API
      * functions
      * Treat unknown XML as anydata */
@@ -1529,7 +1528,9 @@ netconf_module_load(clicon_handle h)
     /* Load restconf collection */
     if (yang_spec_parse_module(h, "ietf-netconf-list-pagination", NULL, yspec)< 0)
 	goto done;    
-
+    /* Load restconf collection */
+    if (yang_spec_parse_module(h, "ietf-netconf-list-pagination", NULL, yspec)< 0)
+	goto done;    
     retval = 0;
  done:
     return retval;